pyproject.toml
[project]
name = "kairon"
version = "0.1.0"
description = "A research and drafting system using Google Gemini"
authors = [
    {name = "Akash",email = "akashkallai27@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.10",
    "tavily-python>=0.3.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "langchain-community>=0.0.10",
    "langchain-google-genai>=0.0.5"
]

[tool.poetry]
name = "kairon"
version = "0.1.0"
description = "A research and drafting system using Google Gemini"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.0"
langchain-google-genai = "^0.0.5"
tavily-python = "^0.3.1"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
